# mysql: Improve MySQL 5.6 GTID parsing performance