--- a/go/mysql/mysql56_gtid_set.go
+++ b/go/mysql/mysql56_gtid_set.go
@@ -70,10 +70,10 @@ func parseInterval(s string) (interval, error) {
 	}
 }
 
-// parseMysql56GTIDSet is registered as a GTIDSet parser.
+// ParseMysql56GTIDSet is registered as a GTIDSet parser.
 //
 // https://dev.mysql.com/doc/refman/5.6/en/replication-gtids-concepts.html
-func parseMysql56GTIDSet(s string) (Mysql56GTIDSet, error) {
+func ParseMysql56GTIDSet(s string) (Mysql56GTIDSet, error) {
 	set := Mysql56GTIDSet{}
 
 	// gtid_set: uuid_set [, uuid_set] ...
@@ -657,7 +657,7 @@ func popInterval(dst *interval, s1, s2 *[]interval) bool {
 
 func init() {
 	gtidSetParsers[Mysql56FlavorID] = func(s string) (GTIDSet, error) {
-		return parseMysql56GTIDSet(s)
+		return ParseMysql56GTIDSet(s)
 	}
 }
 
@@ -665,11 +665,11 @@ func init() {
 // The result is also a string.
 // An error is thrown if parsing is not possible for either GTIDSets
 func Subtract(lhs, rhs string) (string, error) {
-	lhsSet, err := parseMysql56GTIDSet(lhs)
+	lhsSet, err := ParseMysql56GTIDSet(lhs)
 	if err != nil {
 		return "", err
 	}
-	rhsSet, err := parseMysql56GTIDSet(rhs)
+	rhsSet, err := ParseMysql56GTIDSet(rhs)
 	if err != nil {
 		return "", err
 	}