--- a/go/mysql/mysql56_gtid_set_test.go
+++ b/go/mysql/mysql56_gtid_set_test.go
@@ -19,7 +19,6 @@ package mysql
 import (
 	"fmt"
 	"reflect"
-	"sort"
 	"strings"
 	"testing"
 
@@ -43,7 +42,7 @@ func TestSortSIDList(t *testing.T) {
 		{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
 		{1, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
 	}
-	sort.Sort(sidList(input))
+	sortSIDs(input)
 	if !reflect.DeepEqual(input, want) {
 		t.Errorf("got %#v, want %#v", input, want)
 	}