## Description

This is a small optimization pass for MySQL 5.6 GTID Set parsing performance. This is something that we do very often during VStreaming and which we were performing quite inefficiently.

Before/after:

```
name                   old time/op    new time/op    delta
MySQL56GTIDParsing-16    13.0µs ± 1%     6.8µs ± 2%  -47.60%  (p=0.000 n=9+10)

name                   old alloc/op   new alloc/op   delta
MySQL56GTIDParsing-16    4.38kB ± 0%    3.10kB ± 0%  -29.07%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
MySQL56GTIDParsing-16      68.0 ± 0%      25.0 ± 0%  -63.24%  (p=0.000 n=10+10)
```

cc @dbussink 

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [x] "Backport me!" label has been added if this change should be backported
-   [x] Tests were added or are not required
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->